Added several new files.
Several new subsystems, command groups, and commands were added. These
are currently a structure for the loader, flipper, and firing of the
robot. A different approach was also used for the inverted motors as
well (the setMaxOutput parameter was changed to take a negative float).

This code has been tested and currently can drive the robot with a
joystick and a Xbox controller. Other components are still under
testing.